correction on task 1 of function pointers

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -13,7 +13,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	long unsigned int i;
+	unsigned int i;
 
 	i = 0;
